file updates
I set a monthly schedule to try and update portions of my files and the website since I often find it so frustrating when I come across other researcher's websites that are hopelessly out of date (and so helpful when they aren't!). This is my first attempt at trying to keep up with this schedule, which involves some changes to my CV, EDI, and teaching statements. Big changes are:

- Added in my Chinese name (love you mom!). I'm still not sure about whether I'll keep that long term, but I'm starting to do it for publishing so I figured might as well.
- Cleaned up my EDI statement to be shorter and more to the point. 
- Ditto with the teaching statement.

Also debating about whether I want to also add in text related to my personal background and/or mentoring experience since that should probably go somewhere but never quite feels like the right place for it. We'll see how I feel in the future if I can keep this up.