monthly update
The penultimate update before the 2021 job season. This includes a few substantial changes to my CV to cut down the length to 5 pages, which include splitting off both mentees (previously "students") and publications into their own document so I can just list highlights. We'll have to see what the feedback is when I discuss it with people here.

I also updated my teaching statement a bit given the mini-course is currently running (ahhhhh!!!) and to include information on SURP 2021 in my EDI statement. I might include a paragraph on mentorship there next time around as well, but for now I'm happy with it.