appearances update
I realized in my quest to be much more thorough in my own accounting of appearances (mainly in response to how painful it was to gather materials for PTR last time around) I had not added in panel participation or media appearances, and that I had completely forgotten about Astronomy on Tap! (Eek!)

Hopefully, this should be the last update to my documents before the website eventually gets revamped.